+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+### Added
+* Provider name check
+
 ## [0.0.22] - 2021-04-26
 
 ### Added